RAG-Document-Intelligence-Hub/requirements.txt
# Core RAG dependencies
openai>=1.0.0
langchain>=0.1.0
tiktoken>=0.5.0

# Vector stores
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0

# OCR (optional but recommended)
pytesseract>=0.3.10
Pillow>=10.0.0

# Data processing
pandas>=2.0.0

# Text processing
unidecode>=1.3.0

# Web interface
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-magic>=0.4.27
